accessibility_chatbot/requirements.txt
Flask
python-dotenv
google-generativeai
Pillow
PyPDF2
reportlab
markdown